chore(clerk-js, types): Move session verification under SessionResource
