+++ packages/clerk-js/src/core/resources/Session.ts
@@ -1,10 +1,19 @@
 import { runWithExponentialBackOff } from '@clerk/shared';
 import { is4xxError } from '@clerk/shared/error';
 import type {
+  __experimental_SessionVerificationJSON,
+  __experimental_SessionVerificationResource,
+  __experimental_SessionVerifyAttemptFirstFactorParams,
+  __experimental_SessionVerifyAttemptSecondFactorParams,
+  __experimental_SessionVerifyCreateParams,
+  __experimental_SessionVerifyPrepareFirstFactorParams,
+  __experimental_SessionVerifyPrepareSecondFactorParams,
   ActJWTClaim,
   CheckAuthorization,
+  EmailCodeConfig,
   GetToken,
   GetTokenOptions,
+  PhoneCodeConfig,
   SessionJSON,
   SessionResource,
   SessionStatus,
@@ -13,9 +22,11 @@ import type {
 } from '@clerk/types';
 
 import { unixEpochToDate } from '../../utils/date';
+import { clerkInvalidStrategy } from '../errors';
 import { eventBus, events } from '../events';
 import { SessionTokenCache } from '../tokenCache';
 import { BaseResource, PublicUserData, Token, User } from './internal';
+import { SessionVerification } from './SessionVerification';
 
 export class Session extends BaseResource implements SessionResource {
   pathRoot = '/client/sessions';
@@ -122,6 +133,98 @@ export class Session extends BaseResource implements SessionResource {
       typeof organizationId === 'undefined' ? this.lastActiveOrganizationId : organizationId;
     return [this.id, template, resolvedOrganizationId, this.updatedAt.getTime()].filter(Boolean).join('-');
   }
+
+  __experimental_startVerification = async ({
+    level,
+    maxAge,
+  }: __experimental_SessionVerifyCreateParams): Promise<__experimental_SessionVerificationResource> => {
+    const json = (
+      await BaseResource._fetch({
+        method: 'POST',
+        path: `/client/sessions/${this.id}/verify`,
+        body: {
+          level,
+          maxAge,
+        } as any,
+      })
+    )?.response as unknown as __experimental_SessionVerificationJSON;
+
+    return new SessionVerification(json);
+  };
+
+  __experimental_prepareFirstFactorVerification = async (
+    factor: __experimental_SessionVerifyPrepareFirstFactorParams,
+  ): Promise<__experimental_SessionVerificationResource> => {
+    let config;
+    switch (factor.strategy) {
+      case 'email_code':
+        config = { emailAddressId: factor.emailAddressId } as EmailCodeConfig;
+        break;
+      case 'phone_code':
+        config = {
+          phoneNumberId: factor.phoneNumberId,
+          default: factor.default,
+        } as PhoneCodeConfig;
+        break;
+      default:
+        clerkInvalidStrategy('Session.prepareFirstFactorVerification', (factor as any).strategy);
+    }
+
+    const json = (
+      await BaseResource._fetch({
+        method: 'POST',
+        path: `/client/sessions/${this.id}/verify/prepare_first_factor`,
+        body: {
+          ...config,
+          strategy: factor.strategy,
+        } as any,
+      })
+    )?.response as unknown as __experimental_SessionVerificationJSON;
+
+    return new SessionVerification(json);
+  };
+
+  __experimental_attemptFirstFactorVerification = async (
+    attemptFactor: __experimental_SessionVerifyAttemptFirstFactorParams,
+  ): Promise<__experimental_SessionVerificationResource> => {
+    const json = (
+      await BaseResource._fetch({
+        method: 'POST',
+        path: `/client/sessions/${this.id}/verify/attempt_first_factor`,
+        body: { ...attemptFactor, strategy: attemptFactor.strategy } as any,
+      })
+    )?.response as unknown as __experimental_SessionVerificationJSON;
+
+    return new SessionVerification(json);
+  };
+
+  __experimental_prepareSecondFactorVerification = async (
+    params: __experimental_SessionVerifyPrepareSecondFactorParams,
+  ): Promise<__experimental_SessionVerificationResource> => {
+    const json = (
+      await BaseResource._fetch({
+        method: 'POST',
+        path: `/client/sessions/${this.id}/verify/prepare_second_factor`,
+        body: params as any,
+      })
+    )?.response as unknown as __experimental_SessionVerificationJSON;
+
+    return new SessionVerification(json);
+  };
+
+  __experimental_attemptSecondFactorVerification = async (
+    params: __experimental_SessionVerifyAttemptSecondFactorParams,
+  ): Promise<__experimental_SessionVerificationResource> => {
+    const json = (
+      await BaseResource._fetch({
+        method: 'POST',
+        path: `/client/sessions/${this.id}/verify/attempt_second_factor`,
+        body: params as any,
+      })
+    )?.response as unknown as __experimental_SessionVerificationJSON;
+
+    return new SessionVerification(json);
+  };
 
   protected fromJSON(data: SessionJSON | null): this {
     if (!data) {
